Add PDO

+++ pages/admin/_users.php
@@ -7,37 +7,47 @@ if (!defined('PSWeb') || PSWeb !== true) { Header('Location: /404'); return; }
 <div class="silver-bk"><div class="clr"></div>	
 <?PHP
 # Редактирование пользователя
-if(isset($_GET["edit"])){
-$eid = intval($_GET["edit"]);
-$db->Query("SELECT *, INET_NTOA(db_users_a.ip) uip FROM db_users_a, db_users_b WHERE db_users_a.id = db_users_b.id AND db_users_b.id = '$eid' LIMIT 1");
+if(isset($_GET['edit'])){
+$eid = intval($_GET['edit']);
+$result = $pdo->prepare("SELECT *, INET_NTOA(db_users_a.ip) uip FROM db_users_a, db_users_b WHERE db_users_a.id = db_users_b.id AND db_users_b.id = :eid LIMIT 1");
+$result->execute(array('eid'=>$eid));
 # Проверяем на существование
-if($db->NumRows() != 1){ echo "<center><b>Указанный пользователь не найден</b></center><BR />"; }
+if($result->rowCount() != 1){ echo '<center><b>Указанный пользователь не найден</b></center><BR />'; }
 # Добавляем дерево
-if(isset($_POST["set_tree"])){
-$tree = $_POST["set_tree"];
-$type = ($_POST["type"] == 1) ? "-1" : "+1";
-	$db->Query("UPDATE db_users_b SET {$tree} = {$tree} {$type} WHERE id = '$eid'");
-	$db->Query("SELECT *, INET_NTOA(db_users_a.ip) uip FROM db_users_a, db_users_b WHERE db_users_a.id = db_users_b.id AND db_users_b.id = '$eid' LIMIT 1");
-	echo "<center><b>Дерево добавлено</b></center><BR />";
+if(isset($_POST['set_tree'])){
+    $tree = $_POST['set_tree'];
+    if(array_key_exists($tree,$items)){
+        $type = ($_POST['type'] == 1) ? "-1" : "+1";
+        $result=$pdo->prepare("UPDATE db_users_b SET {$tree} = {$tree} {$type} WHERE id = :eid");
+        $result->execute(array('eid'=>$eid));
+        $result=$pdo->prepare("SELECT *, INET_NTOA(db_users_a.ip) uip FROM db_users_a, db_users_b WHERE db_users_a.id = db_users_b.id AND db_users_b.id = :eid LIMIT 1");
+        $result->execute(array('eid'=>$eid));
+        echo '<center><b>Дерево добавлено</b></center><BR />';
+    }else{
+        echo '<center><b>Что-то не так</b></center><BR />';
+    }
 }
 # Пополняем баланс
-if(isset($_POST["balance_set"])){
-$sum = intval($_POST["sum"]);
-$bal = $_POST["schet"];
-$type = ($_POST["balance_set"] == 1) ? "-" : "+";
-$string = ($type == "-") ? "У пользователя снято {$sum} серебра" : "Пользователю добавлено {$sum} серебра";
-	$db->Query("UPDATE db_users_b SET {$bal} = {$bal} {$type} {$sum} WHERE id = '$eid'");
-	$db->Query("SELECT *, INET_NTOA(db_users_a.ip) uip FROM db_users_a, db_users_b WHERE db_users_a.id = db_users_b.id AND db_users_b.id = '$eid' LIMIT 1");
-	echo "<center><b>$string</b></center><BR />";	
+if(isset($_POST['balance_set'])){
+    $sum = intval($_POST['sum']);
+    $bal = $_POST['schet'];
+    $type = ($_POST['balance_set'] == 1) ? "-" : "+";
+    $string = ($type == "-") ? "У пользователя снято {$sum} серебра" : "Пользователю добавлено {$sum} серебра";
+    $result=$pdo->prepare("UPDATE db_users_b SET {$bal} = {$bal} {$type} {$sum} WHERE id = :eid");
+    $result->execute(array('eid'=>$eid));
+    $result=$pdo->prepare("SELECT *, INET_NTOA(db_users_a.ip) uip FROM db_users_a, db_users_b WHERE db_users_a.id = db_users_b.id AND db_users_b.id = :eid LIMIT 1");
+    $result->execute(array('eid'=>$eid));
+    echo "<center><b>$string</b></center><BR />";	
 }
 # Забанить пользователя
-if(isset($_POST["banned"])){
-	$db->Query("UPDATE db_users_a SET banned = '".intval($_POST["banned"])."' WHERE id = '$eid'");
-	$db->Query("SELECT *, INET_NTOA(db_users_a.ip) uip FROM db_users_a, db_users_b WHERE db_users_a.id = db_users_b.id AND db_users_b.id = '$eid' LIMIT 1");
-	echo "<center><b>Пользователь ".($_POST["banned"] > 0 ? "забанен" : "разбанен")."</b></center><BR />";
-	
+if(isset($_POST['banned'])){
+    $db->Query("UPDATE db_users_a SET banned = :banned WHERE id = :eid");
+    $result->execute(array('eid'=>$eid,'banned'=>intval($_POST['banned'])));
+    $result=$pdo->prepare("SELECT *, INET_NTOA(db_users_a.ip) uip FROM db_users_a, db_users_b WHERE db_users_a.id = db_users_b.id AND db_users_b.id = :eid LIMIT 1");
+    $result->execute(array('eid'=>$eid));
+    echo '<center><b>Пользователь '.($_POST["banned"] > 0 ? 'забанен' : 'разбанен').'</b></center><BR />';
 }
-$data = $db->FetchArray();
+$data = $result->fetch();
 ?>
 <table width="100%" border="0">
   <tr bgcolor="#efefef">
@@ -52,12 +62,6 @@ $data = $db->FetchArray();
     <td style="padding-left:10px;">Email:</td>
     <td width="200" align="center"><?=$data["email"]; ?></td>
   </tr>
-  <tr>
-    <td style="padding-left:10px;">Пароль:</td>
-    <td width="200" align="center"><?=$data["pass"]; ?></td>
-  </tr>
-  
-  
   <tr bgcolor="#efefef">
     <td style="padding-left:10px;">Серебро (Покупки):</td>
     <td width="200" align="center"><?=sprintf("%.2f",$data["money_b"]); ?></td>
@@ -117,8 +121,9 @@ $data = $db->FetchArray();
     <td style="padding-left:10px;">Рефералов:</td>
 	
 	<?PHP
-	$db->Query("SELECT COUNT(*) FROM db_users_a WHERE referer_id = '".$data["id"]."'");
-	$counter_res = $db->FetchRow();
+	$result=$pdo->prepare("SELECT COUNT(*) FROM db_users_a WHERE referer_id = :user_id");
+        $result->execute(array('user_id'=>$data['id']));
+	$counter_res = $result->fetchColumn();
 	?>
 	
     <td width="200" align="center"><?=$data["referals"]; ?> [<?=$counter_res; ?>] чел.</td>
@@ -203,7 +208,7 @@ return;
 }
 
 ?>
-<form action="/?menu=admin4ik&sel=users&search" method="post">
+<form action="/admin/users/search" method="post">
 <table width="250" border="0" align="center">
   <tr>
     <td><b>Логин:</b></td>
@@ -216,25 +221,29 @@ return;
 <?PHP
 
 function sort_b($int_s){
-	$int_s = intval($int_s);
-	switch($int_s){
-		case 1: return "db_users_a.user";
-		case 2: return "all_serebro";
-		case 3: return "all_trees";
-		case 4: return "db_users_a.date_reg";
-		default: return "db_users_a.id";
-	}
+    $int_s = intval($int_s);
+    switch($int_s){
+        case 1: return "db_users_a.user";
+        case 2: return "all_serebro";
+        case 3: return "all_trees";
+        case 4: return "db_users_a.date_reg";
+        default: return "db_users_a.id";
+    }
 }
-$sort_b = (isset($_GET["sort"])) ? intval($_GET["sort"]) : 0;
+$sort_b = (isset($_GET['sort'])) ? intval($_GET['sort']) : 0;
 $str_sort = sort_b($sort_b);
-$num_p = (isset($_GET["page"]) AND intval($_GET["page"]) < 1000 AND intval($_GET["page"]) >= 1) ? (intval($_GET["page"]) -1) : 0;
+$num_p = (isset($_GET['page']) AND intval($_GET['page']) < 1000 AND intval($_GET['page']) >= 1) ? (intval($_GET['page']) -1) : 0;
 $lim = $num_p * 100;
-if(isset($_GET["search"])){
-$search = $_POST["sear"];
-$db->Query("SELECT *, (db_users_b.a_t + db_users_b.b_t + db_users_b.c_t + db_users_b.d_t + db_users_b.e_t) all_trees, (db_users_b.money_b + db_users_b.money_p) all_serebro 
-FROM db_users_a, db_users_b WHERE db_users_a.id = db_users_b.id AND db_users_a.user = '$search' ORDER BY {$str_sort} DESC LIMIT {$lim}, 100");
-}else $db->Query("SELECT *, (db_users_b.a_t + db_users_b.b_t + db_users_b.c_t + db_users_b.d_t + db_users_b.e_t) all_trees, (db_users_b.money_b + db_users_b.money_p) all_serebro FROM db_users_a, db_users_b WHERE db_users_a.id = db_users_b.id ORDER BY {$str_sort} DESC LIMIT {$lim}, 100");
-if($db->NumRows() > 0){
+if(isset($_GET['search'])){
+    $search = $_POST["sear"];
+    $result=$pdo->prepare("SELECT *, (db_users_b.a_t + db_users_b.b_t + db_users_b.c_t + db_users_b.d_t + db_users_b.e_t) all_trees, (db_users_b.money_b + db_users_b.money_p) all_serebro FROM db_users_a, db_users_b WHERE db_users_a.id = db_users_b.id AND db_users_a.user = :user ORDER BY {$str_sort} DESC LIMIT {$lim}, 100");
+    $result->execute(array(
+        'user'=>$search
+    ));
+}else{
+    $result=$pdo->query("SELECT *, (db_users_b.a_t + db_users_b.b_t + db_users_b.c_t + db_users_b.d_t + db_users_b.e_t) all_trees, (db_users_b.money_b + db_users_b.money_p) all_serebro FROM db_users_a, db_users_b WHERE db_users_a.id = db_users_b.id ORDER BY {$str_sort} DESC LIMIT {$lim}, 100");
+}
+if($result->rowCount() > 0){
 ?>
 <table cellpadding='3' cellspacing='0' border='0' bordercolor='#336633' align='center' width="99%">
   <tr bgcolor="#efefef">
@@ -245,29 +254,24 @@ if($db->NumRows() > 0){
 	<td align="center" width="100" class="m-tb"><a href="/admin/users/sort/4" class="stn-sort">Зарегистрирован</a></td>
   </tr>
 <?PHP
-
-	while($data = $db->FetchArray()){
-	
-	?>
-	<tr class="htt">
-    <td align="center"><?=$data["id"]; ?></td>
-    <td align="center"><a href="/admin/users/edit/<?=$data["id"]; ?>" class="stn"><?=$data["user"]; ?></a></td>
-    <td align="center"><?=sprintf("%.2f",$data["all_serebro"]); ?></td>
+    while($data = $result->fetch()){
+    ?>
+    <tr class="htt">
+        <td align="center"><?=$data["id"]; ?></td>
+        <td align="center"><a href="/admin/users/edit/<?=$data["id"]; ?>" class="stn"><?=$data["user"]; ?></a></td>
+        <td align="center"><?=sprintf("%.2f",$data["all_serebro"]); ?></td>
 	<td align="center"><?=$data["all_trees"]; ?></td>
 	<td align="center"><?=date("d.m.Y",$data["date_reg"]); ?></td>
-  	</tr>
+    </tr>
 	<?PHP
-	
 	}
-
 ?>
-
 </table>
 <BR />
 <?PHP
-
-
-}else echo "<center><b>На данной странице нет записей</b></center><BR />";
+}else{
+    echo '<center><b>На данной странице нет записей</b></center><BR />';
+}
 	if(isset($_GET["search"])){	
 	?>
 	</div>
@@ -275,14 +279,14 @@ if($db->NumRows() > 0){
 	<?PHP
 		return;
 	}
-$db->Query("SELECT COUNT(*) FROM db_users_a");
-$all_pages = $db->FetchRow();
-	if($all_pages > 100){
-	$sort_b = (isset($_GET["sort"])) ? intval($_GET["sort"]) : 0;
-	$nav = new navigator;
-	$page = (isset($_GET["page"]) AND intval($_GET["page"]) < 1000 AND intval($_GET["page"]) >= 1) ? (intval($_GET["page"])) : 1;
-	echo "<BR /><center>".$nav->Navigation(10, $page, ceil($all_pages / 100), "/admin/users/sort/{$sort_b}&page="), "</center>";
-	}
+$result=$pdo->query("SELECT COUNT(*) FROM db_users_a");
+$all_pages = $result->fetchColumn();
+if($all_pages > 100){
+    $sort_b = (isset($_GET['sort'])) ? intval($_GET['sort']) : 0;
+    $nav = new navigator;
+    $page = (isset($_GET['page']) AND intval($_GET['page']) < 1000 AND intval($_GET['page']) >= 1) ? (intval($_GET["page"])) : 1;
+    echo "<BR /><center>".$nav->Navigation(10, $page, ceil($all_pages / 100), "/admin/users/sort/{$sort_b}/page/"), "</center>";
+}
 ?>
 </div>
 <div class="clr"></div>
